# api.yaml
openapi: 3.0.1
info:
  title: Base
  description: Description
  version: 1.0.0
servers:
  - url: 'https://api-ecommerce.hakaya.technology'
  - url: 'http://localhost:3000'
paths:
...
  /login:
    post:
      tags:
        - Public
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        200:
          $ref: '#/components/responses/id'
  /verify:
    post:
      tags:
        - Public
      summary: Verify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/verify'

      responses:
        200:
          $ref: '#/components/responses/userLoggedIn'
...
  /cities:
    post:
      tags:
        - City
      summary: add City
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/countrySchema'
      responses:
        200:
          $ref: '#/components/responses/country'
    get:
      tags:
        - City
      summary: get Cityies
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/country'
  /cities/{id}:
    put:
      tags:
        - City
      summary: update City
      parameters:
        - name: id
          in: path
          required: true
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/countrySchema'
      responses:
        200:
          $ref: '#/components/responses/country'

    get:
      tags:
        - City
      summary: get City
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/country'

    delete:
      tags:
        - City
      summary: delete City
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/message'
...
  /cities/{cityId}/areas:
    post:
      tags:
        - Area
      summary: Add Area
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/citySchema'
      responses:
        200:
          $ref: '#/components/responses/city'
    get:
      tags:
        - City
      summary: get City
      parameters:
        - name: countryId
          in: path
          required: true
          schema:
            type: string

      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/city'
  /cities/{cityId}/areas/{areaId}:
    put:
      tags:
        - Area
      summary: update Area
      security:
        - ApiKeyAuth: []
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string

        - name: areaId
          in: path
          required: true
          schema:
            type: string

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/citySchema'
      responses:
        200:
          $ref: '#/components/responses/city'

    get:
      tags:
        - Area
      summary: Get Area details
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string

        - name: areaId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/city'

    delete:
      tags:
        - Area
      summary: Delete Area
      parameters:
        - name: cityId
          in: path
          required: true
          schema:
            type: string

        - name: areaId
          in: path
          required: true
          schema:
            type: string
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/message'
...
  /complaint:
    post:
      tags:
        - Complaint
      summary: add Complaint
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/complaintSchema'
      responses:
        200:
          $ref: '#/components/responses/complaint'
    get:
      tags:
        - complaint
      summary: get complaints
      security:
        - ApiKeyAuth: []
      responses:
        200:
          $ref: '#/components/responses/complaint'
...
  /search:
    post:
      tags:
        - Search
      summary: add Search input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/searchSchema'
      responses:
        200:
          $ref: '#/components/responses/search'
...
  /search/{deviceId}:
    get:
      tags:
        - Search
      summary: get Search inputs by deviceId
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/search'
    delete:
      tags:
        - Search
      summary: Clear Search inputs by deviceId
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          $ref: '#/components/responses/message'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
...
    id:
      description: Login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/idSchema'
...
    country:
      description: Login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/countrySchema'
...
    city:
      description: Login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/citySchema'
...
    search:
      description: get search.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/searchResponse'
...
    userLoggedIn:
      description: Login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/userLoggedInSchema'
    message:
      description: Message.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/messageSchema'
...
  schemas:
...
    idSchema:
      type: object
      properties:
        id:
          type: "string"
          example: "5c9b8a228e6a017fdfa2488c"
    userLoggedInSchema:
      type: object
      properties:
        _id:
          type: "string"
          example: "5c9b8a228e6a017fdfa2488c"
        mobile:
          type: "string"
          example: "+2011242"
        roles:
          type: "array"
          items:
            type: "string"
          example:
            - "customer"
        token:
          type: "string"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1YzliOGEyMjhlNmEwMTdmZGZhMjQ4OGMiLCJtb2JpbGUiOiIrMjAxMTI0MiIsInJvbGVzIjpbImN1c3RvbWVyIl0sImlhdCI6MTU1MzY5NzMxNH0.XsVW_qqBRnD5mplYVutE-dRZt2eCDhEr8E21UKtIRB4"
    login:
      type: object
      properties:
        mobile:
          type: "string"
          example: "+2011242"
        type:
          type: "string"
          example: "customer"
    verify:
      type: object
      properties:
        user:
          type: "string"
          example: "+2011242"
        mobile_token:
          type: "number"
          example: 332859
    messageSchema:
      type: object
      properties:
        message:
          type: "string"
          example: "Item has been deleted"
...
    countrySchema:
      type: "object"
      properties:
        name:
          type: "string"
        translation:
          type: "object"
          properties:
            ar:
              type: "object"
              properties:
                name:
                  type: "string"
...
    citySchema:
      type: "object"
      properties:
        name:
          type: "string"
        city:
          type: "string"
        translation:
          type: "object"
          properties:
            ar:
              type: "object"
              properties:
                name:
                  type: "string"
...
    searchSchema:
      type: object
      properties:
        device_id:
          type: "string"
          example: "Item1"
        search_input:
          type: "string"
...
    searchResponse:
      type: object
      properties:
        device_id:
          type: "string"
        search_inputs:
          type: "array"
          items:
            type: "string"
...